refactor: create action component now as upsert component
