+++ NERBABO.Frontend/src/app/features/actions/upsert-actions/upsert-actions.component.html
@@ -1,5 +1,7 @@
 <div class="modal-header">
-  <h4 class="modal-title">Registar Ação Formação</h4>
+  <h4 class="modal-title">
+    {{ isUpdate ? "Atualizar" : "Registar" }} Ação Formação
+  </h4>
   <button
     type="button"
     class="btn-close"
